refactor: move files to correct destination
